security_tester/requirements.txt
flask==2.3.3
python-owasp-zap-v2.4==0.0.20
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
Werkzeug==2.3.7